lookups.py: disable local caching of Simbad objects.
Looking for local Subjects before looking for Simbad is going to
generate broken data. E.g., if we have M80 and M81 in our local
cache, and the user searches for M8, autoSuggest will only propose
M80 and M81, and be happy with it. More subjects might be available
in Simbad's database.