Make IOTD Candidates selection a bit fairer
Now, the 25 most liked images are selected, then 25 random images (still
in the 7 day window, as before). They are randomized and then presented
without clues that can identify the author of the image (of course if
there's a source watermark, there's nothing we can do).

This should remove some bias that the person selecting the IOTD might
have.