(infra) Make pipeline use django-storage's cache-busting
This avoids needing to manually set the MEDIA_VERSION. To enable this, I had to remve references to missing images in third party CSS files.

To really take advantage of this, #496 must be solved, or it won't affect files referenced using a path leading with {{STATIC_URL}}.

This is part of #585.