(ui) Remove "iconic" icon set
This gets rid of 1379 files that were contributing significantly in
making "collectstatic" slow on S3.

Additionally, this patch removed a piece of code that was only run for
users with no JS support in their browser, and that used the removed
icons. A lot of AstroBin doesn't work without JS, so we're not going
to start now to support something so obsolete.

Closes #602