(subscription) Do not update or delete a subscription that is not current
This because Stripe creates a subscription object also if the payment fails, and sets it to expire within 24 hours if a payment is not made. However, if a user starts a new checkout session, and succeeds in creating a subscription, that's the subscription that will be deleted in 24 hours! Stripe supports multiple user subscriptions of the same kind, while we don't.